STAT-5350-Translator/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
filelock==3.19.1
fonttools==4.60.2
fpdf2==2.8.4
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.90
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
sympy==1.14.0
tesseract==0.1.3
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.1
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
